Move repo create logic to an interface in the image context
