+++ pkg/image/context.go
@@ -24,6 +24,10 @@ type rpmResolver interface {
 	Resolve(packages *Packages, localPackagesPath, outputDir string) (rpmDirPath string, pkgList []string, err error)
 }
 
+type rpmRepoCreator interface {
+	Create(path string) error
+}
+
 type Context struct {
 	// ImageConfigDir is the root directory storing all configuration files.
 	ImageConfigDir string
@@ -38,5 +42,6 @@ type Context struct {
 	KubernetesScriptInstaller    kubernetesScriptInstaller
 	KubernetesArtefactDownloader kubernetesArtefactDownloader
 	// RPMResolver responsible for resolving rpm/package dependencies
-	RPMResolver rpmResolver
+	RPMResolver    rpmResolver
+	RPMRepoCreator rpmRepoCreator
 }
